Setup notebook to run (one at a time)
I tweaked thresholds and set up the ipynb.
Putting each of these material properties together will require setting the thresholds as a variable, settting up the distribution graph to look pretty with all separate variables (I doubt it can be done automatically), and making a dictionary connecting the path/title to a prettier name for figures.